HealthBar, UIWind

+++ Assets/Scripts/Interfaces/IDamage.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamage
 {
-    void Damage(float damage);
+    void Damage(int damage);
 }
